pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=0.0.24",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.1",
    "tavily-python>=0.5.1",
]
